+++ src/components/popup/popup.ts
@@ -14,15 +14,23 @@ import type { CSSResultGroup } from 'lit';
  *  operations in your listener or consider debouncing it.
  *
  * @slot - The popup's content.
- * @slot anchor - The element the popup will be anchored to.
+ * @slot anchor - The element the popup will be anchored to. If the anchor lives outside of the popup, you can use the
+ *  `anchor` attribute or property instead.
  *
  * @csspart arrow - The arrow's container. Avoid setting `top|bottom|left|right` properties, as these values are
  *  assigned dynamically as the popup moves. This is most useful for applying a background color to match the popup, and
  *  maybe a border or box shadow.
  * @csspart popup - The popup's container. Useful for setting a background color, box shadow, etc.
  *
- * @cssproperty [--arrow-size=4px] - The size of the arrow. Note that an arrow won't be shown unless the `arrow` attribute is used.
+ * @cssproperty [--arrow-size=4px] - The size of the arrow. Note that an arrow won't be shown unless the `arrow`
+ *  attribute is used.
  * @cssproperty [--arrow-color=var(--sl-color-neutral-0)] - The color of the arrow.
+ * @cssproperty [--auto-size-available-width] - A read-only custom property that determines the amount of width the
+ *  popup can be before overflowing. Useful for positioning child elements that need to overflow. This property is only
+ *  available when using `auto-size`.
+ * @cssproperty [--auto-size-available-height] - A read-only custom property that determines the amount of height the
+ *  popup can be before overflowing. Useful for positioning child elements that need to overflow. This property is only
+ *  available when using `auto-size`.
  */
 @customElement('sl-popup')
 export default class SlPopup extends LitElement {
@@ -32,8 +40,14 @@ export default class SlPopup extends LitElement {
   @query('.popup') public popup: HTMLElement;
   @query('.popup__arrow') private arrowEl: HTMLElement;
 
-  private anchor: HTMLElement | null;
+  private anchorEl: HTMLElement | null;
   private cleanup: ReturnType<typeof autoUpdate> | undefined;
+
+  /**
+   * The element the popup will be anchored to. If the anchor lives outside of the popup, you can provide its `id` or a
+   * reference to it here. If the anchor lives inside the popup, use the `anchor` slot instead.
+   */
+  @property() anchor: Element | string;
 
   /**
    * Activates the positioning logic and shows the popup. When this attribute is removed, the positioning logic is torn
@@ -82,7 +96,7 @@ export default class SlPopup extends LitElement {
    * The amount of padding between the arrow and the edges of the popup. If the popup has a border-radius, for example,
    * this will prevent it from overflowing the corners.
    */
-  @property({ type: Number }) arrowPadding = 10;
+  @property({ attribute: 'arrow-padding', type: Number }) arrowPadding = 10;
 
   /**
    * When set, placement of the popup will flip to the opposite site to keep it in view. You can use
@@ -127,11 +141,7 @@ export default class SlPopup extends LitElement {
   @property({ type: Object }) flipBoundary: Element | Element[];
 
   /** The amount of padding, in pixels, to exceed before the flip behavior will occur. */
-  @property({
-    attribute: 'flip-padding',
-    type: Number
-  })
-  flipPadding = 0;
+  @property({ attribute: 'flip-padding', type: Number }) flipPadding = 0;
 
   /** Moves the popup along the axis to keep it in view when clipped. */
   @property({ type: Boolean }) shift = false;
@@ -144,11 +154,7 @@ export default class SlPopup extends LitElement {
   @property({ type: Object }) shiftBoundary: Element | Element[];
 
   /** The amount of padding, in pixels, to exceed before the shift behavior will occur. */
-  @property({
-    attribute: 'shift-padding',
-    type: Number
-  })
-  shiftPadding = 0;
+  @property({ attribute: 'shift-padding', type: Number }) shiftPadding = 0;
 
   /** When set, this will cause the popup to automatically resize itself to prevent it from overflowing. */
   @property({ attribute: 'auto-size', type: Boolean }) autoSize = false;
@@ -161,11 +167,7 @@ export default class SlPopup extends LitElement {
   @property({ type: Object }) autoSizeBoundary: Element | Element[];
 
   /** The amount of padding, in pixels, to exceed before the auto-size behavior will occur. */
-  @property({
-    attribute: 'auto-size-padding',
-    type: Number
-  })
-  autoSizePadding = 0;
+  @property({ attribute: 'auto-size-padding', type: Number }) autoSizePadding = 0;
 
   async connectedCallback() {
     super.connectedCallback();
@@ -179,19 +181,31 @@ export default class SlPopup extends LitElement {
     this.stop();
   }
 
-  async handleAnchorSlotChange() {
+  async handleAnchorChange() {
     await this.stop();
 
-    this.anchor = this.querySelector<HTMLElement>('[slot="anchor"]');
+    if (this.anchor && typeof this.anchor === 'string') {
+      // Locate the anchor by id
+      const root = this.getRootNode() as Document | ShadowRoot;
+      this.anchorEl = root.getElementById(this.anchor);
+    } else if (this.anchor instanceof HTMLElement) {
+      // Use the anchor's reference
+      this.anchorEl = this.anchor;
+    } else {
+      // Look for a slotted anchor
+      this.anchorEl = this.querySelector<HTMLElement>('[slot="anchor"]');
+    }
 
     // If the anchor is a <slot>, we'll use the first assigned element as the target since slots use `display: contents`
     // and positioning can't be calculated on them
-    if (this.anchor instanceof HTMLSlotElement) {
-      this.anchor = this.anchor.assignedElements({ flatten: true })[0] as HTMLElement;
+    if (this.anchorEl instanceof HTMLSlotElement) {
+      this.anchorEl = this.anchorEl.assignedElements({ flatten: true })[0] as HTMLElement;
     }
 
-    if (!this.anchor) {
-      throw new Error('Invalid anchor element: no child with slot="anchor" was found.');
+    if (!this.anchorEl) {
+      throw new Error(
+        'Invalid anchor element: no anchor could be found using the anchor slot or the anchor attribute.'
+      );
     }
 
     this.start();
@@ -199,11 +213,11 @@ export default class SlPopup extends LitElement {
 
   private start() {
     // We can't start the positioner without an anchor
-    if (!this.anchor) {
+    if (!this.anchorEl) {
       return;
     }
 
-    this.cleanup = autoUpdate(this.anchor, this.popup, () => {
+    this.cleanup = autoUpdate(this.anchorEl, this.popup, () => {
       this.reposition();
     });
   }
@@ -214,6 +228,8 @@ export default class SlPopup extends LitElement {
         this.cleanup();
         this.cleanup = undefined;
         this.removeAttribute('data-current-placement');
+        this.style.removeProperty('--auto-size-available-width');
+        this.style.removeProperty('--auto-size-available-height');
         requestAnimationFrame(() => resolve());
       } else {
         resolve();
@@ -224,26 +240,31 @@ export default class SlPopup extends LitElement {
   async updated(changedProps: Map<string, unknown>) {
     super.updated(changedProps);
 
+    // Start or stop the positioner when active changes
     if (changedProps.has('active')) {
-      // Start or stop the positioner when active changes
       if (this.active) {
         this.start();
       } else {
         this.stop();
       }
-    } else {
-      // All other properties will trigger a reposition when active
-      if (this.active) {
-        await this.updateComplete;
-        this.reposition();
-      }
+    }
+
+    // Update the anchor when anchor changes
+    if (changedProps.has('anchor')) {
+      this.handleAnchorChange();
+    }
+
+    // All other properties will trigger a reposition when active
+    if (this.active) {
+      await this.updateComplete;
+      this.reposition();
     }
   }
 
   /** Recalculate and repositions the popup. */
   reposition() {
     // Nothing to do if the popup is inactive or the anchor doesn't exist
-    if (!this.active || !this.anchor) {
+    if (!this.active || !this.anchorEl) {
       return;
     }
 
@@ -255,27 +276,7 @@ export default class SlPopup extends LitElement {
       offset({ mainAxis: this.distance, crossAxis: this.skidding })
     ];
 
-    // First, we adjust the size as needed
-    if (this.autoSize) {
-      middleware.push(
-        size({
-          boundary: this.autoSizeBoundary,
-          padding: this.autoSizePadding,
-          apply: ({ availableWidth, availableHeight }) => {
-            // Ensure the panel stays within the viewport when we have lots of menu items
-            Object.assign(this.popup.style, {
-              maxWidth: `${availableWidth}px`,
-              maxHeight: `${availableHeight}px`
-            });
-          }
-        })
-      );
-    } else {
-      // Unset max-width/max-height when we're no longer using this middleware
-      Object.assign(this.popup.style, { maxWidth: '', maxHeight: '' });
-    }
-
-    // Then we flip, as needed
+    // First we flip
     if (this.flip) {
       middleware.push(
         flip({
@@ -288,7 +289,7 @@ export default class SlPopup extends LitElement {
       );
     }
 
-    // Then we shift, as needed
+    // Then we shift
     if (this.shift) {
       middleware.push(
         shift({
@@ -296,6 +297,24 @@ export default class SlPopup extends LitElement {
           padding: this.shiftPadding
         })
       );
+    }
+
+    // Now we adjust the size as needed
+    if (this.autoSize) {
+      middleware.push(
+        size({
+          boundary: this.autoSizeBoundary,
+          padding: this.autoSizePadding,
+          apply: ({ availableWidth, availableHeight }) => {
+            this.style.setProperty('--auto-size-available-width', `${availableWidth}px`);
+            this.style.setProperty('--auto-size-available-height', `${availableHeight}px`);
+          }
+        })
+      );
+    } else {
+      // Cleanup styles if we're no longer using auto-size
+      this.style.removeProperty('--auto-size-available-width');
+      this.style.removeProperty('--auto-size-available-height');
     }
 
     // Finally, we add an arrow
@@ -308,7 +327,7 @@ export default class SlPopup extends LitElement {
       );
     }
 
-    computePosition(this.anchor, this.popup, {
+    computePosition(this.anchorEl, this.popup, {
       placement: this.placement,
       middleware,
       strategy: this.strategy
@@ -341,7 +360,7 @@ export default class SlPopup extends LitElement {
 
   render() {
     return html`
-      <slot name="anchor" @slotchange=${this.handleAnchorSlotChange}></slot>
+      <slot name="anchor" @slotchange=${this.handleAnchorChange}></slot>
 
       <div
         part="popup"