+++ src/components/spinner-circle/spinner-circle.test.ts
@@ -16,7 +16,7 @@ describe('<sl-spinner-circle>', () => {
     it('should have a role of "status".', () => {
       // https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Live_Regions
       const base = el.shadowRoot!.querySelector('[part="base"]')!;
-      expect(base).have.attribute('role', 'status');
+      expect(base).have.attribute('role', 'progessbar');
     });
   });
 });