+++ src/react/radio-button/index.ts
@@ -4,6 +4,5 @@ import Component from '../../components/radio-button/radio-button';
 
 export default createComponent(React, 'sl-radio-button', Component, {
   onSlBlur: 'sl-blur',
-  onSlChange: 'sl-change',
   onSlFocus: 'sl-focus'
 });